restructured utils folder
