+++ index.ts
@@ -17,4 +17,4 @@
 export * from './src/infrastructure/infrastructure';
 export * from './src/model/model';
 export * from './src/service/service';
-export * from './src/utility/utility';
+export * from './src/core/utils/utility';
