+++ src/core/utils/utility.ts
