+++ test/core/utils/TransactionMapping.spec.ts
@@ -18,44 +18,44 @@ import {deepEqual} from 'assert';
 import { expect } from 'chai';
 import { sha3_256 } from 'js-sha3';
 import { convert } from 'nem2-library';
-import { Account } from '../../src/model/account/Account';
-import { Address } from '../../src/model/account/Address';
-import { PropertyModificationType } from '../../src/model/account/PropertyModificationType';
-import { PropertyType } from '../../src/model/account/PropertyType';
-import { PublicAccount } from '../../src/model/account/PublicAccount';
-import { NetworkType } from '../../src/model/blockchain/NetworkType';
-import { MosaicId } from '../../src/model/mosaic/MosaicId';
-import { MosaicNonce } from '../../src/model/mosaic/MosaicNonce';
-import { MosaicProperties } from '../../src/model/mosaic/MosaicProperties';
-import { MosaicSupplyType } from '../../src/model/mosaic/MosaicSupplyType';
-import { NetworkCurrencyMosaic } from '../../src/model/mosaic/NetworkCurrencyMosaic';
-import { AliasActionType } from '../../src/model/namespace/AliasActionType';
-import { NamespaceId } from '../../src/model/namespace/NamespaceId';
-import { NamespaceType } from '../../src/model/namespace/NamespaceType';
-import { AccountLinkTransaction } from '../../src/model/transaction/AccountLinkTransaction';
-import { AccountPropertyTransaction } from '../../src/model/transaction/AccountPropertyTransaction';
-import { AddressAliasTransaction } from '../../src/model/transaction/AddressAliasTransaction';
-import { AggregateTransaction } from '../../src/model/transaction/AggregateTransaction';
-import { Deadline } from '../../src/model/transaction/Deadline';
-import { HashType } from '../../src/model/transaction/HashType';
-import { LinkAction } from '../../src/model/transaction/LinkAction';
-import { LockFundsTransaction } from '../../src/model/transaction/LockFundsTransaction';
-import { ModifyAccountPropertyAddressTransaction } from '../../src/model/transaction/ModifyAccountPropertyAddressTransaction';
-import { ModifyMultisigAccountTransaction } from '../../src/model/transaction/ModifyMultisigAccountTransaction';
-import { MosaicAliasTransaction } from '../../src/model/transaction/MosaicAliasTransaction';
-import { MosaicDefinitionTransaction } from '../../src/model/transaction/MosaicDefinitionTransaction';
-import { MosaicSupplyChangeTransaction } from '../../src/model/transaction/MosaicSupplyChangeTransaction';
-import { MultisigCosignatoryModification } from '../../src/model/transaction/MultisigCosignatoryModification';
-import { MultisigCosignatoryModificationType } from '../../src/model/transaction/MultisigCosignatoryModificationType';
-import { PlainMessage } from '../../src/model/transaction/PlainMessage';
-import { RegisterNamespaceTransaction } from '../../src/model/transaction/RegisterNamespaceTransaction';
-import { SecretLockTransaction } from '../../src/model/transaction/SecretLockTransaction';
-import { SecretProofTransaction } from '../../src/model/transaction/SecretProofTransaction';
-import { TransactionType } from '../../src/model/transaction/TransactionType' ;
-import { TransferTransaction } from '../../src/model/transaction/TransferTransaction';
-import { UInt64 } from '../../src/model/UInt64';
-import { TransactionMapping } from '../../src/utility/TransactionMapping';
-import { TestingAccount } from '../conf/conf.spec';
+import { TransactionMapping } from '../../../src/core/utils/TransactionMapping';
+import { Account } from '../../../src/model/account/Account';
+import { Address } from '../../../src/model/account/Address';
+import { PropertyModificationType } from '../../../src/model/account/PropertyModificationType';
+import { PropertyType } from '../../../src/model/account/PropertyType';
+import { PublicAccount } from '../../../src/model/account/PublicAccount';
+import { NetworkType } from '../../../src/model/blockchain/NetworkType';
+import { MosaicId } from '../../../src/model/mosaic/MosaicId';
+import { MosaicNonce } from '../../../src/model/mosaic/MosaicNonce';
+import { MosaicProperties } from '../../../src/model/mosaic/MosaicProperties';
+import { MosaicSupplyType } from '../../../src/model/mosaic/MosaicSupplyType';
+import { NetworkCurrencyMosaic } from '../../../src/model/mosaic/NetworkCurrencyMosaic';
+import { AliasActionType } from '../../../src/model/namespace/AliasActionType';
+import { NamespaceId } from '../../../src/model/namespace/NamespaceId';
+import { NamespaceType } from '../../../src/model/namespace/NamespaceType';
+import { AccountLinkTransaction } from '../../../src/model/transaction/AccountLinkTransaction';
+import { AccountPropertyTransaction } from '../../../src/model/transaction/AccountPropertyTransaction';
+import { AddressAliasTransaction } from '../../../src/model/transaction/AddressAliasTransaction';
+import { AggregateTransaction } from '../../../src/model/transaction/AggregateTransaction';
+import { Deadline } from '../../../src/model/transaction/Deadline';
+import { HashType } from '../../../src/model/transaction/HashType';
+import { LinkAction } from '../../../src/model/transaction/LinkAction';
+import { LockFundsTransaction } from '../../../src/model/transaction/LockFundsTransaction';
+import { ModifyAccountPropertyAddressTransaction } from '../../../src/model/transaction/ModifyAccountPropertyAddressTransaction';
+import { ModifyMultisigAccountTransaction } from '../../../src/model/transaction/ModifyMultisigAccountTransaction';
+import { MosaicAliasTransaction } from '../../../src/model/transaction/MosaicAliasTransaction';
+import { MosaicDefinitionTransaction } from '../../../src/model/transaction/MosaicDefinitionTransaction';
+import { MosaicSupplyChangeTransaction } from '../../../src/model/transaction/MosaicSupplyChangeTransaction';
+import { MultisigCosignatoryModification } from '../../../src/model/transaction/MultisigCosignatoryModification';
+import { MultisigCosignatoryModificationType } from '../../../src/model/transaction/MultisigCosignatoryModificationType';
+import { PlainMessage } from '../../../src/model/transaction/PlainMessage';
+import { RegisterNamespaceTransaction } from '../../../src/model/transaction/RegisterNamespaceTransaction';
+import { SecretLockTransaction } from '../../../src/model/transaction/SecretLockTransaction';
+import { SecretProofTransaction } from '../../../src/model/transaction/SecretProofTransaction';
+import { TransactionType } from '../../../src/model/transaction/TransactionType' ;
+import { TransferTransaction } from '../../../src/model/transaction/TransferTransaction';
+import { UInt64 } from '../../../src/model/UInt64';
+import { TestingAccount } from '../../conf/conf.spec';
 
 describe('TransactionMapping', () => {
     let account: Account;
